feat(svm-sdk): extract leaf fee base and add regressive fee reader/writer
Extracted SvmLeafFeeReader/Writer abstract base with abstract feeType and
strategyKind properties. Slimmed down linear-fee.ts to thin subclass.
Added SvmRegressiveFeeReader/Writer following the same pattern.
